refactor(ui): Use file based layout for repo layout
The previous implementation used a shared layout in the route file for
repository specific pages and sub-pages that are under `$runIndex`. Use
file based grouping to use TanStack Router more idiomatically.
