+++ ui/src/routes/_layout/organizations/$orgId/products/$productId/repositories/$repoId/-components/evaluator-fields.tsx
@@ -34,7 +34,7 @@ import {
 } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
 import { Switch } from '@/components/ui/switch';
-import { CreateRunFormValues } from '../-create-run-utils';
+import { CreateRunFormValues } from '../_repo-layout/-create-run-utils';
 
 type EvaluatorFieldsProps = {
   form: UseFormReturn<CreateRunFormValues>;
