+++ ui/src/routes/_layout/organizations/$orgId/products/$productId/repositories/$repoId/-components/package-manager-field.tsx
@@ -41,7 +41,10 @@ import { Label } from '@/components/ui/label';
 import { Separator } from '@/components/ui/separator';
 import { cn } from '@/lib/utils';
 import { packageManagers } from '@/routes/_layout/organizations/$orgId/products/$productId/repositories/$repoId/-types';
-import { CreateRunFormValues, PackageManagerId } from '../-create-run-utils';
+import {
+  CreateRunFormValues,
+  PackageManagerId,
+} from '../_repo-layout/-create-run-utils';
 
 type PackageManagerFieldProps = {
   form: UseFormReturn<CreateRunFormValues>;
