+++ ui/src/routes/_layout/organizations/$orgId/products/$productId/repositories/$repoId/_repo-layout/index.tsx
@@ -47,7 +47,7 @@ import {
 } from '@/components/ui/tooltip';
 import { toast } from '@/lib/toast';
 import { paginationSearchParameterSchema } from '@/schemas';
-import { RepositoryRunsTable } from './-components/repository-runs-table';
+import { RepositoryRunsTable } from '../-components/repository-runs-table';
 
 const defaultPageSize = 10;
 
@@ -187,7 +187,7 @@ const RepoComponent = () => {
 };
 
 export const Route = createFileRoute(
-  '/_layout/organizations/$orgId/products/$productId/repositories/$repoId/'
+  '/_layout/organizations/$orgId/products/$productId/repositories/$repoId/_repo-layout/'
 )({
   validateSearch: paginationSearchParameterSchema,
   loaderDeps: ({ search: { page, pageSize } }) => ({ page, pageSize }),
