+++ ui/src/routes/_layout/organizations/$orgId/products/$productId/repositories/$repoId/_repo-layout/users/route.tsx
@@ -20,14 +20,14 @@
 import { createFileRoute, Outlet, redirect } from '@tanstack/react-router';
 
 export const Route = createFileRoute(
-  '/_layout/organizations/$orgId/products/$productId/repositories/$repoId/secrets'
+  '/_layout/organizations/$orgId/products/$productId/repositories/$repoId/_repo-layout/users'
 )({
   component: () => <Outlet />,
   beforeLoad: ({ context, params }) => {
     if (
       !context.auth.hasRole([
         'superuser',
-        `permission_repository_${params.repoId}_write_secrets`,
+        `role_repository_${params.repoId}_admin`,
       ])
     ) {
       throw redirect({
